pyproject.toml
[project]
name = "sciagent"
version = "0.1.0"
description = "AGT paper search and Zotero workflow agent"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "anyio>=4.6.0",
  "httpx>=0.28.0",
  "langgraph>=0.2.0",
  "langsmith",
  "pydantic>=2.6",
  "pydantic-settings>=2.6",
  "pyzotero",
  "responses",
  "semanticscholar",
  "streamlit>=1.40",
  "structlog",
  "tenacity",
  "vcrpy"
]

[project.optional-dependencies]
keywords = [
  "keybert>=0.8.5",
  "sentence-transformers>=3.0.0"
]
spell = [
  "pyspellchecker>=0.8.1"
]
rerank = [
  "sentence-transformers>=3.0.0"
]

[dependency-groups]
dev = [
  "pre-commit",
  "pyright",
  "pytest",
  "pytest-vcr",
  "ruff"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agt"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.uv]
package = true
